+++ src/components/ItemList/ItemList.jsx
@@ -1,240 +1,16 @@
 import React from 'react';
 import ItemItem from '@/components/ItemList/ItemListItem';
 import * as Styled from '@/components/ItemList/ItemList.styled';
-import img from '@/assets/images/apple.jpg';
 
-const dummy = [
-  {
-    id: 1,
-    imageUrl: img,
-    Name: '사탕',
-    option: [
-      {
-        optionId: 1,
-        optionName: '초코맛',
-        optionPrice: 300,
-        optionStock: 10,
-      },
-      {
-        optionId: 2,
-        optionName: '딸기맛',
-        optionPrice: 200,
-        optionStock: 5,
-      },
-    ],
-    price: 100,
-    discountRate: 50,
-    isDisplay: true,
-  },
-  {
-    id: 2,
-    imageUrl: img,
-    Name: '과자',
-    option: [
-      {
-        optionId: 1,
-        optionName: '초코맛',
-        optionPrice: 300,
-        optionStock: 10,
-      },
-      {
-        optionId: 2,
-        optionName: '딸기맛',
-        optionPrice: 200,
-        optionStock: 5,
-      },
-    ],
-    price: 100,
-    discountRate: 0,
-    isDisplay: true,
-  },
-  {
-    id: 3,
-    imageUrl: img,
-    Name: '아이스크림',
-    option: [
-      {
-        optionId: 1,
-        optionName: '초코맛',
-        optionPrice: 300,
-        optionStock: 10,
-      },
-      {
-        optionId: 2,
-        optionName: '딸기맛',
-        optionPrice: 200,
-        optionStock: 5,
-      },
-    ],
-    price: 100,
-    discountRate: 15,
-    isDisplay: true,
-  },
-  {
-    id: 4,
-    imageUrl: img,
-    Name: '아이스크림',
-    option: [
-      {
-        optionId: 1,
-        optionName: '초코맛',
-        optionPrice: 300,
-        optionStock: 10,
-      },
-      {
-        optionId: 2,
-        optionName: '딸기맛',
-        optionPrice: 200,
-        optionStock: 5,
-      },
-    ],
-    price: 100,
-    discountRate: 15,
-    isDisplay: true,
-  },
-  {
-    id: 5,
-    imageUrl: img,
-    Name: '아이스크림',
-    option: [
-      {
-        optionId: 1,
-        optionName: '초코맛',
-        optionPrice: 300,
-        optionStock: 10,
-      },
-      {
-        optionId: 2,
-        optionName: '딸기맛',
-        optionPrice: 200,
-        optionStock: 5,
-      },
-    ],
-    price: 100,
-    discountRate: 15,
-    isDisplay: true,
-  },
-  {
-    id: 6,
-    imageUrl: img,
-    Name: '아이스크림',
-    option: [
-      {
-        optionId: 1,
-        optionName: '초코맛',
-        optionPrice: 300,
-        optionStock: 10,
-      },
-      {
-        optionId: 2,
-        optionName: '딸기맛',
-        optionPrice: 200,
-        optionStock: 5,
-      },
-    ],
-    price: 100,
-    discountRate: 15,
-    isDisplay: true,
-  },
-  {
-    id: 7,
-    imageUrl: img,
-    Name: '아이스크림',
-    option: [
-      {
-        optionId: 1,
-        optionName: '초코맛',
-        optionPrice: 300,
-        optionStock: 10,
-      },
-      {
-        optionId: 2,
-        optionName: '딸기맛',
-        optionPrice: 200,
-        optionStock: 5,
-      },
-    ],
-    price: 100,
-    discountRate: 15,
-    isDisplay: true,
-  },
-  {
-    id: 8,
-    imageUrl: img,
-    Name: '아이스크림',
-    option: [
-      {
-        optionId: 1,
-        optionName: '초코맛',
-        optionPrice: 300,
-        optionStock: 10,
-      },
-      {
-        optionId: 2,
-        optionName: '딸기맛',
-        optionPrice: 200,
-        optionStock: 5,
-      },
-    ],
-    price: 100,
-    discountRate: 15,
-    isDisplay: true,
-  },
-  {
-    id: 9,
-    imageUrl: img,
-    Name: '아이스크림',
-    option: [
-      {
-        optionId: 1,
-        optionName: '초코맛',
-        optionPrice: 300,
-        optionStock: 10,
-      },
-      {
-        optionId: 2,
-        optionName: '딸기맛',
-        optionPrice: 200,
-        optionStock: 5,
-      },
-    ],
-    price: 100,
-    discountRate: 15,
-    isDisplay: true,
-  },
-  {
-    id: 10,
-    imageUrl: img,
-    Name: '아이스크림',
-    option: [
-      {
-        optionId: 1,
-        optionName: '초코맛',
-        optionPrice: 300,
-        optionStock: 10,
-      },
-      {
-        optionId: 2,
-        optionName: '딸기맛',
-        optionPrice: 200,
-        optionStock: 5,
-      },
-    ],
-    price: 100,
-    discountRate: 15,
-    isDisplay: true,
-  },
-];
-
-const ItemList = () => {
-  const filteredDummy = dummy.filter(item => item.isDisplay === true);
-
+const ItemList = ({ data }) => {
   return (
-    <Styled.ItemList>
-      {filteredDummy.map(item => {
-        return <ItemItem key={`${item.id + item.Name}`} item={item} />;
-      })}
-    </Styled.ItemList>
+    <Styled.ItemListSection>
+      <Styled.ItemList>
+        {data.map(item => {
+          return <ItemItem key={`${item.id + item.Name}`} item={item} />;
+        })}
+      </Styled.ItemList>
+    </Styled.ItemListSection>
   );
 };
